refactor session client to separate file/module
