+++ packages/agw-client/src/abstractClient.ts
@@ -5,19 +5,15 @@ import {
   createClient,
   createPublicClient,
   createWalletClient,
-  custom,
   http,
   type Transport,
 } from 'viem';
 import { type ChainEIP712 } from 'viem/zksync';
 
-import type { SessionConfig } from './sessions.js';
 import { getSmartAccountAddressFromInitialSigner } from './utils.js';
 import {
   type AbstractWalletActions,
   globalWalletActions,
-  type SessionClientActions,
-  sessionWalletActions,
 } from './walletActions.js';
 
 /**
@@ -48,21 +44,10 @@ interface CreateAbstractClientParameters {
   isPrivyCrossApp?: boolean;
 }
 
-interface CreateSessionClientParameters {
-  account: Address;
-  chain: ChainEIP712;
-  signer: Account;
-  session: SessionConfig;
-  transport?: Transport;
-}
-
 type AbstractClientActions = AbstractWalletActions<ChainEIP712, Account>;
 
 export type AbstractClient = Client<Transport, ChainEIP712, Account> &
   AbstractClientActions;
-
-export type SessionClient = Client<Transport, ChainEIP712, Account> &
-  SessionClientActions;
 
 export async function createAbstractClient({
   signer,
@@ -104,59 +89,4 @@ export async function createAbstractClient({
     globalWalletActions(signerWalletClient, publicClient, isPrivyCrossApp),
   );
   return abstractClient as AbstractClient;
-}
-
-export interface ToSessionClientParams {
-  client: AbstractClient;
-  signer: Account;
-  session: SessionConfig;
-}
-
-export function toSessionClient({
-  client,
-  signer,
-  session,
-}: ToSessionClientParams) {
-  return createSessionClient({
-    account: client.account.address,
-    chain: client.chain,
-    session,
-    signer,
-    transport: custom(client.transport),
-  });
-}
-
-export function createSessionClient({
-  account,
-  signer,
-  chain,
-  transport,
-  session,
-}: CreateSessionClientParameters) {
-  if (!transport) {
-    transport = http();
-  }
-
-  const publicClient = createPublicClient({
-    transport: http(),
-    chain,
-  });
-
-  const baseClient = createClient({
-    account,
-    chain: chain,
-    transport,
-  });
-
-  const signerWalletClient = createWalletClient({
-    account: signer,
-    chain: chain,
-    transport,
-  });
-
-  const sessionClient = baseClient.extend(
-    sessionWalletActions(signerWalletClient, publicClient, session) as any, // why do we need to cast to any here?
-  );
-
-  return sessionClient as SessionClient;
 }
